+++ tests/e2e/api/src/repositories/rest/products/index.ts
@@ -1,7 +1,9 @@
 import { createProductTransformer } from './shared';
 import { simpleProductRESTRepository } from './simple-product';
+import { externalProductRESTRepository } from './external-product';
 
 export {
 	createProductTransformer,
+	externalProductRESTRepository,
 	simpleProductRESTRepository,
 };